+++ examples/reverse.rs
@@ -11,8 +11,8 @@ fn reverse<T: Clone>(xs: &[T]) -> Vec<T> {
 }
 
 fn main() {
-    fn equality_after_applying_twice(xs: Vec<int>) -> bool {
+    fn equality_after_applying_twice(xs: Vec<isize>) -> bool {
         xs == reverse(reverse(xs.as_slice()).as_slice())
     }
-    quickcheck(equality_after_applying_twice as fn(Vec<int>) -> bool);
+    quickcheck(equality_after_applying_twice as fn(Vec<isize>) -> bool);
 }
